+++ polls/tests/tests_polls_dates.py
@@ -1,11 +1,8 @@
+"""Django test for Django Model."""
 from django.test import TestCase
-
 # Create your tests here.
 import datetime
-
 from django.utils import timezone
-from django.urls import reverse
-
 from polls.models import Question
 
 
@@ -109,85 +106,3 @@ class QuestionModelTests(TestCase):
         ended_poll = Question(pub_date=time, end_date=timezone.now()
                               - datetime.timedelta(days=1))
         self.assertIs(ended_poll.can_vote(), False)
-
-
-class QuestionIndexViewTests(TestCase):
-    """Tests for index page."""
-
-    def test_no_questions(self):
-        """If no questions exist, an appropriate message is displayed."""
-        response = self.client.get(reverse('polls:index'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_past_question(self):
-        """
-        Question with a pub date in the past are displayed on the
-        index page.
-        """
-        create_question(question_text="Past question.", start=-30, end=-29)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_future_question(self):
-        """
-        Questions with a pub_date in the future aren't displayed on
-        the index page.
-        """
-        create_question(question_text="Future question.", start=30, end=35)
-        response = self.client.get(reverse('polls:index'))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_future_question_and_past_question(self):
-        """
-        Even if both past and future questions exist, only past questions
-        are displayed.
-        """
-        create_question(question_text="Past question.", start=-30, end=-25)
-        create_question(question_text="Future question.", start=30, end=35)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_two_past_questions(self):
-        """The questions index page may display multiple questions."""
-        create_question(question_text="Past question 1.", start=-30, end=-25)
-        create_question(question_text="Past question 2.", start=-5, end=-3)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question 2.>', '<Question: Past question 1.>']
-        )
-
-
-class QuestionDetailViewTests(TestCase):
-    """Tests for detail page"""
-
-    def test_future_question(self):
-        """
-        The detail view of a question with a pub_date in the future
-        returns a 404 not found.
-        """
-        future_question = create_question(question_text='Future question.',
-                                          start=5, end=10)
-        url = reverse('polls:detail', args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-
-    def test_past_question(self):
-        """
-        The detail view of a question with a pub_date in the past
-        displays the quesion's text.
-        """
-        past_question = create_question(question_text='Past Question.',
-                                        start=-5, end=-2)
-        url = reverse('polls:detail', args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
